Split Codex SQL session search helper
